Moved Auth Endpoitns to dedicated folder
Also, removed old auth check method
